+++ cosmetics-web/app/views/notification_build/add_import_country.html.slim
@@ -11,7 +11,7 @@
         = form_for(@notification, url: wizard_path, method: :put) do |form|
             .govuk-form-group
                 = form.label :import_country, "Country", class: "govuk-label"
-                = form.text_field :import_country,
-                    class: ["govuk-input", import_country_errors ? "govuk-input--error" : nil]
+                = form.select :import_country, @countries, { include_blank: "" },
+                    class: "govuk-select", id: "location-autocomplete"
             .govuk-form-group
                 = form.submit "Continue", class: "govuk-button"
