+++ shared-web/app/assets/javascripts/location_picker.js
