+++ shared-web/app/assets/javascripts/main.js
@@ -6,3 +6,4 @@
 //= require govuk-country-and-territory-autocomplete/dist/location-autocomplete.min.js
 //= require accessible-autocomplete/dist/accessible-autocomplete.min.js
 //= require govuk-frontend/all.js
+//= require_tree .
