+++ shared-web/app/assets/stylesheets/shared-imports.scss
@@ -3,3 +3,4 @@ $govuk-image-url-function: 'image-url';
 $govuk-font-url-function: 'font-url';
 @import "govuk-frontend/all";
 @import "navbar";
+@import "autocomplete"
